api/requirements.txt
asgiref==3.9.1
cfgv==3.4.0
click==8.2.1
cryptography==45.0.2
distlib==0.4.0
Django==5.2.5
django-cors-headers==4.7.0
djangorestframework==3.15.2
filelock==3.19.1
gunicorn==21.2.0
identify==2.6.13
nodeenv==1.9.1
packaging==25.0
platformdirs==4.3.8
pre_commit==4.3.0
PyJWT==2.10.1
PyYAML==6.0.2
redis==5.0.1
requests==2.32.3
rq==1.16.2
ruff==0.8.4
sqlparse==0.5.3
ua-parser==1.0.1
ua-parser-builtins==0.18.0.post1
user-agents==2.2.0
virtualenv==20.34.0
